+++ quickparcel/quickparcel/src/main/java/edu/cit/lariosa/quickparcel/repository/AdminRepository.java
@@ -1,0 +1,7 @@
+package edu.cit.lariosa.quickparcel.repository;
+
+import edu.cit.lariosa.quickparcel.entity.Admin;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AdminRepository extends JpaRepository<Admin, Long> {
+}