+++ quickparcel/quickparcel/src/main/java/edu/cit/lariosa/quickparcel/repository/ParcelRepository.java
@@ -1,0 +1,9 @@
+package edu.cit.lariosa.quickparcel.repository;
+
+import edu.cit.lariosa.quickparcel.entity.Parcel;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
+
+public interface ParcelRepository extends JpaRepository<Parcel, Long> {
+    List<Parcel> findBySenderId(Long senderId);
+}