+++ quickparcel/quickparcel/src/main/java/edu/cit/lariosa/quickparcel/repository/RiderRepository.java
@@ -1,0 +1,7 @@
+package edu.cit.lariosa.quickparcel.repository;
+
+import edu.cit.lariosa.quickparcel.entity.Rider;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RiderRepository extends JpaRepository<Rider, Long> {
+}