+++ quickparcel/quickparcel/src/main/java/edu/cit/lariosa/quickparcel/repository/SenderRepository.java
@@ -1,0 +1,7 @@
+package edu.cit.lariosa.quickparcel.repository;
+
+import edu.cit.lariosa.quickparcel.entity.Sender;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SenderRepository extends JpaRepository<Sender, Long> {
+}